Checking in before wholesale change to f_size int->size_t

+++ cps_pp/include/alg/enum_int.h
@@ -19,7 +19,7 @@ typedef double Float;
 
 typedef double IFloat;
 
-typedef uint64_t Pointer;
+typedef uint32_t Pointer;
 
 enum DirType {
 	DIR_X = 0,
